Refactor: Update Quran database and data handling
