+++ faith_data/src/androidMain/kotlin/net/thechance/mena/faith/data/database/QuranDatabaseBuilder.android.kt
@@ -8,8 +8,8 @@ actual class QuranDatabaseBuilder(private val context: Context) {
         context = context.applicationContext,
         name = DATABASE_NAME
     ).fallbackToDestructiveMigration(false)
-        .createFromAsset(databaseUri.removePrefix(prefix))
+        .createFromAsset(databaseUri.removePrefix(pathPrefix))
 
-    internal actual val prefix: String
+    internal actual val pathPrefix: String
         get() = "file:///android_asset/"
 }
