+++ faith_data/src/commonMain/kotlin/net/thechance/mena/faith/data/database/AyahDto.kt
@@ -2,23 +2,19 @@ package net.thechance.mena.faith.data.database
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
-import androidx.room.Index
 import androidx.room.PrimaryKey
 
-@Entity(
-    tableName = "ayat",
-    indices = [Index(value = ["sura_no"])]
-)
+@Entity(tableName = "ayat")
 data class AyahDto(
-    @PrimaryKey val id: Int?,
+    @PrimaryKey val id: Int,
     @ColumnInfo(name = "sura_no") val surahNumber: Int,
     @ColumnInfo(name = "sura_name_en") val surahName: String,
     @ColumnInfo(name = "sura_name_ar") val surahNameAr: String,
     @ColumnInfo(name = "aya_no") val number: Int,
     @ColumnInfo(name = "aya_text") val text: String,
     @ColumnInfo(name = "aya_text_emlaey") val ayaTextEmlaey: String,
-    @ColumnInfo(name = "jozz") val juzz: Int,
-    @ColumnInfo(name = "page") val page: Int,
     @ColumnInfo(name = "line_start") val lineStart: Int,
     @ColumnInfo(name = "line_end") val lineEnd: Int,
+    val jozz: Int,
+    val page: Int,
 )
