+++ faith_data/src/commonMain/kotlin/net/thechance/mena/faith/data/database/QuranDatabaseBuilder.kt
@@ -4,7 +4,7 @@ import androidx.room.RoomDatabase
 import net.thechance.mena.faith_data.Res
 
 expect class QuranDatabaseBuilder {
-    internal val prefix:String
+    internal val pathPrefix:String
     fun getBuilder(): RoomDatabase.Builder<QuranDatabase>
 }
