--- faith_data/src/commonMain/kotlin/net/thechance/mena/faith/data/database/SimpleAyahDto.kt
@@ -1,11 +1,0 @@
-package net.thechance.mena.faith.data.database
-
-import androidx.room.ColumnInfo
-
-data class SimpleAyahDto(
-val id: Int?,
-@ColumnInfo(name = "aya_no") val number: Int?,
-@ColumnInfo(name = "aya_text") val text: String?,
-    @ColumnInfo(name = "aya_text_emlaey") val ayaTextEmlaey: String?,
-)
-
