+++ faith_data/src/commonMain/kotlin/net/thechance/mena/faith/data/di/quranDatabaseModule.kt
@@ -2,4 +2,4 @@ package net.thechance.mena.faith.data.di
 
 import org.koin.core.module.Module
 
-expect fun faithPlatformModule(): Module
+expect fun quranDatabaseModule(): Module
