+++ faith_data/src/commonMain/kotlin/net/thechance/mena/faith/data/mapper/SurahMapper.kt
@@ -1,0 +1,15 @@
+package net.thechance.mena.faith.data.mapper
+
+import net.thechance.mena.faith.data.database.SurahDto
+import net.thechance.mena.faith.domain.entity.Surah
+
+fun SurahDto.toSurah(): Surah {
+    val surahOrder = Surah.SurahOrder.entries.first { it.order == order }
+    return Surah(
+        id = order,
+        order = surahOrder,
+        name = nameEn,
+        ayahCount = ayahCount,
+        isMakkia = surahOrder.isMakkia
+    )
+}
