--- faith_data/src/commonMain/kotlin/net/thechance/mena/faith/data/utils/AppDispatchers.kt
@@ -1,10 +1,0 @@
-package net.thechance.mena.faith.data.utils
-
-import kotlinx.coroutines.CoroutineDispatcher
-
-expect class DataDispatchers() {
-    val main: CoroutineDispatcher
-    val io: CoroutineDispatcher
-    val default: CoroutineDispatcher
-    val unconfined: CoroutineDispatcher
-}
